move depenetrate to plugin
remove outdated TODO comment
